+++ core/src/main/java/org/sterl/spring/persistent_tasks/history/repository/TriggerHistoryDetailRepository.java
@@ -2,33 +2,29 @@ package org.sterl.spring.persistent_tasks.history.repository;
 
 import java.util.List;
 
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
-import org.sterl.spring.persistent_tasks.api.HistoryOverview;
+import org.sterl.spring.persistent_tasks.api.TaskHistoryOverview;
 import org.sterl.spring.persistent_tasks.history.model.TriggerHistoryDetailEntity;
 
 public interface TriggerHistoryDetailRepository extends HistoryTriggerRepository<TriggerHistoryDetailEntity> {
 
     @Query("""
-            SELECT new org.sterl.spring.persistent_tasks.api.HistoryOverview(
-                     e.instanceId,
+            SELECT new org.sterl.spring.persistent_tasks.api.TaskHistoryOverview(
                      e.data.key.taskName,
                      count(1) as entryCount,
-                     MIN(e.data.start) as start,
-                     MAX(e.data.end) as end,
-                     MIN(e.data.createdTime) as createdTime,
-                     MAX(e.data.executionCount) as executionCount,
-                     AVG(e.data.runningDurationInMs) as runningDurationInMs
+                     MIN(e.data.runAt) as firstRun,
+                     MAX(e.data.runAt) as lastRun,
+                     MAX(e.data.runningDurationInMs) as maxDuration,
+                     MIN(e.data.runningDurationInMs) as minDuration,
+                     AVG(e.data.runningDurationInMs) as avgDuration
                    )
              FROM #{#entityName} e
-             GROUP BY 
-                 e.instanceId,
-                 e.data.key.taskName
-             ORDER BY end DESC, createdTime DESC
+             WHERE e.data.end IS NOT NULL
+             GROUP BY e.data.key.taskName
+             ORDER BY e.data.key.taskName ASC
              """)
-     Page<HistoryOverview> listHistoryOverview(Pageable page);
+     List<TaskHistoryOverview> listTaskHistoryOverview();
 
      @Query("""
              SELECT e