+++ core/src/test/java/org/sterl/spring/persistent_tasks/history/repository/TriggerHistoryDetailRepositoryTest.java
@@ -6,7 +6,6 @@ import java.time.OffsetDateTime;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.PageRequest;
 import org.sterl.spring.persistent_tasks.AbstractSpringTest;
 import org.sterl.spring.persistent_tasks.api.TriggerStatus;
 import org.sterl.spring.persistent_tasks.history.model.TriggerHistoryDetailEntity;
@@ -18,9 +17,9 @@ class TriggerHistoryDetailRepositoryTest extends AbstractSpringTest {
     @Test
     void testGrouping() {
         // GIVEN
-        var history1 = newHistoryEntry(TriggerStatus.RUNNING, OffsetDateTime.now());
+        var history1 = newHistoryEntry(TriggerStatus.FAILED, OffsetDateTime.now());
         subject.save(history1);
-        var history2 = newHistoryEntry(TriggerStatus.RUNNING, OffsetDateTime.now());
+        var history2 = newHistoryEntry(TriggerStatus.SUCCESS, OffsetDateTime.now());
         history2.setInstanceId(history1.getInstanceId());
         history2.getData().setKey(history1.getKey());
         subject.save(history2);
@@ -28,10 +27,10 @@ class TriggerHistoryDetailRepositoryTest extends AbstractSpringTest {
         subject.save(newHistoryEntry(TriggerStatus.RUNNING, OffsetDateTime.now()));
 
         // WHEN
-        var result = subject.listHistoryOverview(PageRequest.of(0, 10));
+        var result = subject.listTaskHistoryOverview();
         
         // THEN
-        assertThat(result.getTotalElements()).isEqualTo(2L);
+        assertThat(result.size()).isEqualTo(2L);
     }
 
     private TriggerHistoryDetailEntity newHistoryEntry(TriggerStatus s, OffsetDateTime created) {
@@ -39,6 +38,8 @@ class TriggerHistoryDetailRepositoryTest extends AbstractSpringTest {
         history.setId(null);
         history.setCreatedTime(created);
         history.getData().setStatus(s);
+        history.getData().setStart(OffsetDateTime.now());
+        history.getData().setEnd(OffsetDateTime.now().plusSeconds(10));
         return history;
     }